+++ src/components/app/app.test.js
@@ -2,42 +2,75 @@ import React from "react";
 import renderer from "react-test-renderer";
 import App from "./app.jsx";
 
+const details = {
+  bgPoster: `BG_HREF`,
+  cover: `POSTER_HREF`,
+  genre: `comedy`,
+  year: 2001,
+  rate: 8.8,
+  votes: 666,
+  director: `Kevin Smith`,
+  actors: [`actor1`, `actor2`],
+  description: {
+    prescription: `prescription`,
+    postscription: `postscription`,
+  },
+};
+
 const FILMS = [
   {
+    id: 0,
     title: `title`,
     src: `src`,
+    details,
   },
   {
+    id: 0,
     title: `title`,
     src: `src`,
+    details,
   },
   {
+    id: 0,
     title: `title`,
     src: `src`,
+    details,
   },
   {
+    id: 0,
     title: `title`,
     src: `src`,
+    details,
   },
   {
+    id: 0,
     title: `title`,
     src: `src`,
+    details,
   },
   {
+    id: 0,
     title: `title`,
     src: `src`,
+    details,
   },
   {
+    id: 0,
     title: `title`,
     src: `src`,
+    details,
   },
   {
+    id: 0,
     title: `title`,
     src: `src`,
+    details,
   },
   {
+    id: 0,
     title: `title`,
     src: `src`,
+    details,
   },
 ];
